デバッグログの削除

+++ db/seeds.rb
@@ -1,13 +1,4 @@
-  # This file should ensure the existence of records required to run the application in every environment (production,
-  # development, test). The code here should be idempotent so that it can be executed at any point in every environment.
-  # The data can then be loaded with the bin/rails db:seed command (or created alongside the database with db:setup).
-  #
-  # Example:
-  #
-  #   ["Action", "Comedy", "Drama", "Horror"].each do |genre_name|
-  #     MovieGenre.find_or_create_by!(name: genre_name)
-  #   end
-  # unless Rails.env.test?
+  return if Rails.env.test?
   # 既存のデータを削除する（開発環境のみ）※削除の順番に注意！
   if Rails.env.development?
     FacilityMatch.destroy_all  # 中間テーブル削除を追加
